LayoutTests/http/tests/referrer-policy-img/no-referrer-when-downgrade/cross-origin-http.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/referrer-policy-img/no-referrer-when-downgrade/cross-origin-http.https.html was allowed to display insecure content from http://127.0.0.1:8000/referrer-policy/resources/image.py?expected=.

Tests the behavior of no-referrer-when-downgrade referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
